requirements.txt
Flask
gunicorn==20.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
matplotlib==3.5.3
numpy==1.23.3
pandas==1.4.4
scikit-learn==1.1.2
scipy==1.9.1
tensorboard==2.10.0
tensorflow-cpu==2.10.0
tensorflow-hub==0.12.0